+++ src/query_all.ts
@@ -4,23 +4,23 @@ import { Maybe, maybe } from '@fluss/core/maybe';
  * Select elements from parent node. By default selects from **document**.
  * Functional variant of `document.querySelectorAll` method.
  */
-export function querySelectorAll<T extends keyof HTMLElementTagNameMap>(
+export function queryAll<T extends keyof HTMLElementTagNameMap>(
   selector: T,
   parent?: ParentNode | Maybe<ParentNode> | null
 ): ReadonlyArray<HTMLElementTagNameMap[T]>;
-export function querySelectorAll<T extends Element>(
+export function queryAll<T extends Element>(
   selector: string,
   parent?: ParentNode | Maybe<ParentNode> | null
 ): ReadonlyArray<T>;
-export function querySelectorAll<T extends Element>(
+export function queryAll<T extends Element>(
   selector: string,
   parent?: ParentNode | Maybe<ParentNode> | null
 ): ReadonlyArray<T> {
-  const array: Array<T> = [];
-  maybe(parent ?? document).map((parentElement) =>
-    parentElement
-      .querySelectorAll<T>(selector)
-      .forEach((item) => array.push(item))
+  return (
+    maybe(parent ?? document)
+      .map((parentElement) => parentElement.querySelectorAll<T>(selector))
+      .map<Array<T>>(Array.from)
+      .map<ReadonlyArray<T>>(Object.freeze)
+      .extract() ?? []
   );
-  return Object.freeze(array);
 }